# docs/openapi.yaml
openapi: 3.0.3
info:
  title: travelmap API
  version: unreleased
  description: >
    The OpenAPI contract for travelmap's API — paths, schemas, headers,
    and status codes for the endpoints actually implemented. That
    includes travelmap's own extensions, at paths outside `/api/v1`,
    alongside the Dawarich-compatible subset. See `docs/api-notes.md`
    for what travelmap's API is as a whole and the reasoning behind
    this contract.


    Upstream's own OpenAPI document
    (`https://raw.githubusercontent.com/Freika/dawarich/master/swagger/v1/swagger.yaml`)
    is the compatibility source of truth for the Dawarich-compatible
    part.
servers:
  - url: /
security:
  - ApiKeyQuery: []
  - BearerAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      summary: Exchange an email and password for an API key
      description: >
        The only way a client that starts with just a password gets the
        API key every other endpoint wants. Always answers 200 with the
        key on success; this server has no two-factor authentication to
        answer 202 with a challenge for.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          $ref: "#/components/responses/LoginOK"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/LoginFailed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/users/me:
    get:
      summary: The account the request authenticated as
      description: >
        What a client calls to confirm that the API key it was configured
        with belongs to somebody.
      operationId: usersMe
      responses:
        "200":
          $ref: "#/components/responses/UsersMeOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    head:
      summary: The account the request authenticated as (headers only)
      description: >
        Clients probe a GET route with HEAD first and read the pagination
        or presence signal off the headers before fetching the body,
        so every GET route here also answers HEAD with the same status
        and headers.
      operationId: usersMeHead
      responses:
        "200":
          $ref: "#/components/responses/UsersMeOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/settings/mobile:
    get:
      summary: The mobile app's own tracking settings
      operationId: getMobileSettings
      responses:
        "200":
          $ref: "#/components/responses/MobileSettingsOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    patch:
      summary: Update the mobile app's own tracking settings
      description: >
        A field the body carries overwrites the stored value; one it omits
        is left as it was. Accepts both forms the published spec's schema
        and example disagree on: flat fields at the top level, or the same
        fields wrapped in a settings object.
      operationId: patchMobileSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MobileSettingsUpdateRequest"
      responses:
        "200":
          $ref: "#/components/responses/MobileSettingsUpdated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/InvalidMobileSettings"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/settings:
    get:
      summary: The map and route-drawing settings
      operationId: getSettings
      responses:
        "200":
          $ref: "#/components/responses/SettingsOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    patch:
      summary: Update the map and route-drawing settings
      description: >
        A field the body carries overwrites the stored value; one it omits
        is left as it was. Accepts both forms the published spec's schema
        and example disagree on, the same contradiction settings/mobile's
        has. immich_url, immich_api_key, photoprism_url and
        photoprism_api_key are accepted without failing the request, but
        stored nowhere — Immich/Photoprism integration is a non-goal. The
        published spec documents no response body for this one, so it
        answers `{}`.
      operationId: patchSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdateRequest"
      responses:
        "200":
          $ref: "#/components/responses/SettingsUpdated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/points:
    get:
      summary: List the authenticated user's points
      description: >
        Optionally narrowed by start_at/end_at (RFC 3339, a bare
        YYYY-MM-DD date, or a Unix timestamp), paginated by page/per_page,
        ordered newest-first unless order=asc. See docs/api-notes.md for
        where the response shape deliberately diverges from the published
        Dawarich schema.
      operationId: listPoints
      parameters:
        - $ref: "#/components/parameters/StartAt"
        - $ref: "#/components/parameters/EndAt"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - $ref: "#/components/parameters/Order"
      responses:
        "200":
          $ref: "#/components/responses/PointsListed"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    head:
      summary: List the authenticated user's points (headers only)
      description: >
        Identical to GET, with no response body. Clients probe with HEAD
        to read X-Total-Pages before fetching any page at all; see the
        note on GET /api/v1/users/me.
      operationId: listPointsHead
      parameters:
        - $ref: "#/components/parameters/StartAt"
        - $ref: "#/components/parameters/EndAt"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - $ref: "#/components/parameters/Order"
      responses:
        "200":
          $ref: "#/components/responses/PointsListed"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      summary: Record a batch of GPS points
      description: >
        Takes the same GeoJSON batch as POST /api/v1/overland/batches and
        differs only in the success status. A Feature with unusable
        coordinates or an unparseable timestamp is dropped rather than
        failing the whole batch. The response body is not part of the
        compatibility contract with upstream (whose own body no client is
        known to read); `created` is the count of points actually inserted
        after deduplication.
      operationId: createPoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationsRequest"
      responses:
        "200":
          $ref: "#/components/responses/PointsCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/points/{id}:
    parameters:
      - $ref: "#/components/parameters/PointId"
    patch:
      summary: Update a point's coordinates
      description: >
        Overwrites latitude and longitude; the timestamp is never part of
        the request and never changes. id not existing, or belonging to a
        different user, both answer 404 — upstream scopes the lookup to
        the authenticated user's own points and treats the two the same.
      operationId: patchPoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PointUpdateRequest"
      responses:
        "200":
          $ref: "#/components/responses/PointUpdated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/InvalidPoint"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: Delete a point
      description: >
        id not existing, or belonging to a different user, both answer
        404, for the same reason as PATCH on this path.
      operationId: deletePoint
      responses:
        "200":
          $ref: "#/components/responses/PointDeleted"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/points/bulk_destroy:
    delete:
      summary: Delete several points at once
      description: >
        An id in point_ids that does not exist, or belongs to a different
        user, is silently skipped rather than failing the request —
        upstream's own `where(id: point_ids).destroy_all` scoping. count
        is how many were actually deleted, which can be fewer than
        point_ids named.
      operationId: bulkDeletePoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PointsBulkDeleteRequest"
      responses:
        "200":
          $ref: "#/components/responses/PointsBulkDeleted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/NoPointsSelected"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/points/tracked_months:
    get:
      summary: The years and months the user has points recorded for
      description: >
        Read from daily_stats, not points directly. Years most-recent-first;
        within a year, months in calendar order.
      operationId: trackedMonths
      responses:
        "200":
          $ref: "#/components/responses/TrackedMonthsOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    head:
      summary: The years and months the user has points recorded for (headers only)
      description: >
        Identical to GET, with no response body. See the note on
        GET /api/v1/users/me.
      operationId: trackedMonthsHead
      responses:
        "200":
          $ref: "#/components/responses/TrackedMonthsOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/stats:
    get:
      summary: Aggregate distance, point and place counts
      description: >
        Distance, point counts and country/city counts aggregated from
        daily_stats, overall and broken down by year and month. The only
        endpoint using camelCase; see docs/api-notes.md for why its
        distances are whole kilometres rather than a JSON number.
      operationId: stats
      responses:
        "200":
          $ref: "#/components/responses/StatsOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    head:
      summary: Aggregate distance, point and place counts (headers only)
      description: >
        Identical to GET, with no response body. See the note on
        GET /api/v1/users/me.
      operationId: statsHead
      responses:
        "200":
          $ref: "#/components/responses/StatsOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/overland/batches:
    post:
      summary: Record a batch of GPS points (Overland app format)
      description: >
        Identical to POST /api/v1/points, aside from the 201 success
        status and accepting the Overland iOS app's own timestamp layout
        (no colon in the zone offset) in addition to RFC 3339.
      operationId: createOverlandBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationsRequest"
      responses:
        "201":
          $ref: "#/components/responses/OverlandBatchCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/tracks:
    get:
      summary: The authenticated user's tracks, as GeoJSON
      description: >
        A FeatureCollection of LineStrings, one per track — a contiguous
        run of points split from its neighbours by a gap exceeding
        TRAVELMAP_TRACK_BREAK_MINUTES. color, avg_speed and duration are
        always the same shape; dominant_mode and dominant_mode_emoji are
        always null, since this server infers no transport mode.
      operationId: listTracks
      parameters:
        - $ref: "#/components/parameters/StartAt"
        - $ref: "#/components/parameters/EndAt"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          $ref: "#/components/responses/TracksOK"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    head:
      summary: The authenticated user's tracks, as GeoJSON (headers only)
      description: >
        Identical to GET, with no response body. See the note on
        GET /api/v1/users/me.
      operationId: listTracksHead
      parameters:
        - $ref: "#/components/parameters/StartAt"
        - $ref: "#/components/parameters/EndAt"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          $ref: "#/components/responses/TracksOK"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/tracks/{id}:
    parameters:
      - $ref: "#/components/parameters/TrackId"
    get:
      summary: One track, as GeoJSON
      description: >
        A single-Feature FeatureCollection, the same per-feature shape
        GET /api/v1/tracks answers with. id not existing, or belonging to
        a different user, both answer 404.
      operationId: getTrack
      responses:
        "200":
          $ref: "#/components/responses/TrackOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    head:
      summary: One track, as GeoJSON (headers only)
      description: >
        Identical to GET, with no response body. See the note on
        GET /api/v1/users/me.
      operationId: getTrackHead
      responses:
        "200":
          $ref: "#/components/responses/TrackOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/tracks/{track_id}/points:
    parameters:
      - $ref: "#/components/parameters/TrackPointsTrackId"
    get:
      summary: The points belonging to one track
      description: >
        Ordered by timestamp ascending. A narrower shape than
        GET /api/v1/points, matching upstream's own documented schema for
        this endpoint — velocity here is a JSON number, unlike
        GET /api/v1/points' own string. track_id not existing, or
        belonging to a different user, both answer 404.
      operationId: trackPoints
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          $ref: "#/components/responses/TrackPointsOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    head:
      summary: The points belonging to one track (headers only)
      description: >
        Identical to GET, with no response body. See the note on
        GET /api/v1/users/me.
      operationId: trackPointsHead
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          $ref: "#/components/responses/TrackPointsOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /webhooks/foursquare:
    post:
      summary: Swarm (Foursquare) User Push notification
      description: >
        Receives one check-in in real time, pushed by Foursquare's own "Push
        API notifications" as the user checks in on Swarm — travelmap's own
        extension, not a Dawarich endpoint. The secret field is checked
        against foursquare.push_secret; checkin.user.id is looked up against
        foursquare_accounts to find which travelmap user it belongs to, and
        the check-in is stored (or updated in place, if already stored) for
        that user. A Foursquare user id with no linked travelmap account is
        not an error: there is simply nothing to store it against. Every
        response has an empty body.
      operationId: foursquareWebhook
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/FoursquarePush"
      responses:
        "200":
          description: >
            Handled: a check-in was stored, or there was nothing to store
            (no checkin parameter, one that is not JSON, or a Foursquare
            user id with no linked travelmap account). Not an error in any
            of those cases.
        "400":
          description: The request body did not parse as a form at all.
        "401":
          description: The secret form field is missing or does not match.
        "500":
          description: The store could not be read or written.
components:
  securitySchemes:
    ApiKeyQuery:
      type: apiKey
      in: query
      name: api_key
      description: >
        Accepted on every endpoint here, regardless of which one upstream's
        own spec documents it for.
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        `Authorization: Bearer <api_key>`. Accepted on every endpoint here,
        regardless of which one upstream's own spec documents it for.
...
  responses:
...
    LoginOK:
      description: The credentials were valid; here is the account's API key.
      headers:
        X-Dawarich-Response:
          $ref: "#/components/headers/DawarichResponse"
        X-Dawarich-Version:
          $ref: "#/components/headers/DawarichVersion"
      content:
        application/json; charset=utf-8:
          schema:
            $ref: "#/components/schemas/Login"
          example:
            user_id: 1
            email: alice@example.com
            api_key: 0f1e2d3c4b5a69788796a5b4c3d2e1f00f1e2d3c4b5a69788796a5b4c3d2e1f0
            status: active
            plan: pro
            subscription_source: none
            active_until: "9999-12-31T23:59:59Z"
    LoginFailed:
      description: >
        The credentials were refused — wrong password, unknown address, or
        a malformed one. Every way of failing gets this same body; how long
        it takes does not say which addresses have accounts.
      headers:
        X-Dawarich-Response:
          $ref: "#/components/headers/DawarichResponse"
        X-Dawarich-Version:
          $ref: "#/components/headers/DawarichVersion"
      content:
        application/json; charset=utf-8:
          schema:
            $ref: "#/components/schemas/AuthError"
          example:
            error: auth_failed
            message: Invalid email or password
    UsersMeOK:
      description: The account the request authenticated as.
      headers:
        X-Dawarich-Response:
          $ref: "#/components/headers/DawarichResponse"
        X-Dawarich-Version:
          $ref: "#/components/headers/DawarichVersion"
      content:
        application/json; charset=utf-8:
          schema:
            $ref: "#/components/schemas/UserEnvelope"
    MobileSettingsOK:
      description: The mobile app's own tracking settings.
      headers:
        X-Dawarich-Response:
          $ref: "#/components/headers/DawarichResponse"
        X-Dawarich-Version:
          $ref: "#/components/headers/DawarichVersion"
      content:
        application/json; charset=utf-8:
          schema:
            $ref: "#/components/schemas/MobileSettingsEnvelope"
...
    BadRequest:
      description: The request body could not be read as the shape this endpoint expects.
      headers:
        X-Dawarich-Response:
          $ref: "#/components/headers/DawarichResponse"
        X-Dawarich-Version:
          $ref: "#/components/headers/DawarichVersion"
      content:
        application/json; charset=utf-8:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: invalid request body
    Unauthorized:
      description: >
        No request authenticated. The body is empty on purpose — upstream's
        `head :unauthorized`, not an oversight — so a client parsing a 401
        body gets nothing rather than something to trip over.
      headers:
        X-Dawarich-Response:
          $ref: "#/components/headers/DawarichResponse"
        X-Dawarich-Version:
          $ref: "#/components/headers/DawarichVersion"
...
    InternalServerError:
      description: >
        The server could not do what was asked, most commonly because its
        database could not be read.
      headers:
        X-Dawarich-Response:
          $ref: "#/components/headers/DawarichResponse"
        X-Dawarich-Version:
          $ref: "#/components/headers/DawarichVersion"
      content:
        application/json; charset=utf-8:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: internal server error
  schemas:
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
...